requirements.txt
torch
torchvision
streamlit
pillow
opencv-python
numpy
matplotlib
scikit-image
